add save edited list title api call
